pyproject.toml
#################################################
# Project metadata
#################################################

[project]
name = "acrobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings>=2.6.1",
    "pydantic-logfire>=0.0.1",
    "python-telegram-bot>=21.7",
    "libsql-client>=0.3.1",
    "sqlparse>=0.5.2",
    "logfire>=2.5.0",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.4",
]

#################################################
# Mypy config
#################################################

[tool.mypy]
follow_imports = "skip"
ignore_missing_imports = true
warn_no_return = true
warn_unused_ignores = true
allow_untyped_globals = true
allow_redefinition = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = "svc.tests.*"
ignore_errors = true

#################################################
# Ruff config
#################################################

[tool.ruff]
respect-gitignore = true
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes `E` and `F` codes by default
select = ["E", "F", "PT", "C4", "I"]
ignore = ["E501"]

per-file-ignores = {}

[tool.ruff.lint.mccabe]
max-complexity = 10

#################################################
# Pytest config
#################################################

[tool.pytest.ini_options]
addopts = "--strict-markers --maxfail 1 --cov src tests/ --no-header"
markers = """
    integration: mark a test as an integration test.
"""
console_output_style = "progress"
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
